commands now have permissions and whitelist kwargs
both fields allow users to determine who can use that command.

permission can be 'broadcaster', 'moderator', 'subscriber', or 'everyone'. although 'everyone' is pretty useless. defining permission implies that viewers higher up on the hierarchy can use the command. so for example, permission='moderator' implies that both mods and the broadcaster can use it.

whitelist is a list of username strings. this allows the user to define which viewers can use the command by name. for example, whitelist='someviewer' means that only viewers with the name 'someviewer' can use it.

note that when used together, permissions take higher priority over whitelists. for example, permission='moderator', whitelist='someviewer' means that ALL moderators and the broadcaster (defined by permissions) and users with the name 'someviewer' can use it.