Now checks parameters for correct data types
1. For functions that the user can access, we now check if the parameters they're sending in are the data type we're expecting. If they're not, we raise a TypeError.

2. For almost all instances in which we expect the user to send in a list of some variable, we now allow them to send just one instance of said variable. TwitchPy should be smart enough to convert the variable into a string whose only element in that variable.
Ex: if a function expects a list of strings and the user only wants to send in one string, they no longer need to do param=['mystring'], they can just do param='mystring' and TwitchPy will internally convert it into ['mystring']
Note: we don't make sure that they're sending in lists either, we just make sure they're sending in something iterable. in theory, the user should be able to send in sets or tuples of some variable instead of lists and it should still work.

3. Added utilities.py module to house functions that help with the above.

4. In the above changes, we use python 3.8's new walrus operator. Because of this, TwitchPy now requires python 3.8+.